+++ server/src/utils/connectDB.ts
@@ -2,8 +2,8 @@ import dotenv from "dotenv";
 dotenv.config();
 import mongoose from "mongoose";
 
-export default function connectDB(mongoUri: string) {
-  return mongoose.connect(mongoUri, {
+export default async function connectDB(mongoUri: string) {
+  await mongoose.connect(mongoUri, {
     dbName: "social-media-app"
   })
 }
